SingleVM: Test SSH connectivity via External IP
Test the external IP connectivity using SSH. This ensures that
the external IP can be used to reach the instance.

Also enhance the setup scripts to setup a unique ssh key for a
given test run. This can be used in the future to eliminate the
need to have the password setup in the cloud-init.

Signed-off-by: Manohar Castelino <manohar.r.castelino@intel.com>