networking: handle CNCI refresh command
Handle the CNCI refresh payload. When we get a CNCI refresh command,
update the route and arp tables to create the GRE point to multipoint
connections between the CNCIs for a tenant. We create the cnci tunnel
device if neccessary, and make sure all the connections exist and
are still valid. We do not store this information persistently, as
a ConcentratorInstanceAdded message will trigger this command
to be sent.

Signed-off-by: Kristen Carlson Accardi <kristen@linux.intel.com>