WIP: configure dnsmasq to support multiple CNCIs per tenant
Add the ability to forward DNS requests for the tenant network
to other CNCIs along the CNCI tunnel. To do this we need to:

  * Add a domain for the tenant network.
  * Add a listen address on the CNCI tunnel IP endpoint.
  * Add a new server address for each peer CNCI for the tenant domain.

DNS requests for the tenants domain will be forwarded to the other
CNCIs if they cannot be handled by the local CNCI. They will not be
forwarded out the interface assigned by the datacenter.

Signed-off-by: Kristen Carlson Accardi <kristen@linux.intel.com>